+++ app/Http/Controllers/Admin/RiderController.php
@@ -239,9 +239,11 @@ class RiderController extends Controller
                 //remove old reference if used referral code updated
                 if($usedRef->used_referral_code != $user->used_referral_code){
                     $old_reffered = User::where('referral_code', $user->used_referral_code)->first();
-                    $reference = ReferralUser::where('user_id', $old_reffered->id)->where('referral_id', $request->id)->first();
-                    if($reference){
-                        $reference->delete();
+                    if($old_reffered){
+                        $reference = ReferralUser::where('user_id', $old_reffered->id)->where('referral_id', $request->id)->first();
+                        if($reference){
+                            $reference->delete();
+                        }
                     }
                 }
